Feature: 앱 꾸미기 clear
