+++ index.html
@@ -1,11 +1,21 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <meta charset="UTF-8">
-    <title>Demo</title>
-  </head>
-  <body>
-    <div id="app"></div>
-    <script src="/main.js"></script>
-  </body>
+
+<head>
+  <meta charset="UTF-8">
+  <title>Demo</title>
+  <style>
+    body {
+      font-size: 16px;
+      margin: 0;
+      padding: 0;
+    }
+  </style>
+</head>
+
+<body>
+  <div id="app"></div>
+  <script src="/main.js"></script>
+</body>
+
 </html>
