+++ src/App.jsx
@@ -1,3 +1,5 @@
+import React from 'react';
+
 import {
   Switch,
   Route,
@@ -6,6 +8,8 @@ import {
 
 import { useDispatch } from 'react-redux';
 
+import styled from '@emotion/styled';
+
 import HomePage from './HomePage';
 import AboutPage from './AboutPage';
 import LoginPage from './LoginPage';
@@ -13,9 +17,31 @@ import RestaurantsPage from './RestaurantsPage';
 import RestaurantPage from './RestaurantPage';
 import NotFoundPage from './NotFoundPage';
 
-import { setAccessToken } from './actions';
+import { setAccessToken } from './slice';
 
 import { loadItem } from './services/storage';
+
+const Container = styled.div({
+  margin: '0 auto',
+  width: '90%',
+  justifyContent: 'center',
+});
+
+const Header = styled.header({
+  backgroundColor: '#EEE',
+  '& h1': {
+    fontSize: '1.5em',
+    margin: 0,
+    padding: '1em .5em',
+  },
+  '& a': {
+    color: '#555',
+    textDecoration: 'none',
+    '&:hover': {
+      color: '#000',
+    },
+  },
+});
 
 export default function App() {
   const dispatch = useDispatch();
@@ -26,12 +52,12 @@ export default function App() {
   }
 
   return (
-    <div>
-      <header>
+    <Container>
+      <Header>
         <h1>
-          <Link to="/">헤더 영역</Link>
+          <Link to="/">헤더</Link>
         </h1>
-      </header>
+      </Header>
       <Switch>
         <Route exact path="/" component={HomePage} />
         <Route path="/about" component={AboutPage} />
@@ -40,6 +66,6 @@ export default function App() {
         <Route path="/restaurants/:id" component={RestaurantPage} />
         <Route component={NotFoundPage} />
       </Switch>
-    </div>
+    </Container>
   );
 }
