+++ src/HomePage.jsx
@@ -1,15 +1,54 @@
+import React from 'react';
+
 import { Link } from 'react-router-dom';
+
+import styled from '@emotion/styled';
+
+const Title = styled.h2({
+  margin: 0,
+  fontSize: '2em',
+  padding: '.4em 0',
+});
+
+const List = styled.ul({
+  display: 'flex',
+  margin: 0,
+  padding: 0,
+  listStyle: 'none',
+});
+
+const Item = styled.li({
+  marginRight: '1em',
+  '& a': {
+    color: '#333',
+    textDecoration: 'none',
+    '&:hover': {
+      fontWeight: 'bold',
+      color: '#000',
+    },
+  },
+});
 
 export default function HomePage() {
   return (
     <div>
-      <h2>Home</h2>
-      <ul>
-        <li><Link to="/about">About</Link></li>
-        <li><Link to="/login">Log in</Link></li>
-        <li><Link to="/restaurants">Restaurants</Link></li>
-        <li><Link to="/xxx">멸망의 길</Link></li>
-      </ul>
+      <Title>
+        Home
+      </Title>
+      <List>
+        <Item>
+          <Link to="/about">About</Link>
+        </Item>
+        <Item>
+          <Link to="/login">Log in</Link>
+        </Item>
+        <Item>
+          <Link to="/restaurants">Restaurants</Link>
+        </Item>
+        <Item>
+          <Link to="/xxx">멸망의 길</Link>
+        </Item>
+      </List>
     </div>
   );
 }
