+++ src/RegionsContainer.jsx
@@ -1,9 +1,14 @@
+import React from 'react';
+
 import { useDispatch, useSelector } from 'react-redux';
+
+import MenuList from './MenuList';
+import MenuItem from './MenuItem';
 
 import {
   selectRegion,
   loadRestaurants,
-} from './actions';
+} from './slice';
 
 import { get } from './utils';
 
@@ -19,9 +24,12 @@ export default function RegionsContainer() {
   }
 
   return (
-    <ul>
+    <MenuList>
       {regions.map((region) => (
-        <li key={region.id}>
+        <MenuItem
+          key={region.id}
+          active={selectedRegion && region.id === selectedRegion.id}
+        >
           <button
             type="button"
             onClick={() => handleClick(region.id)}
@@ -33,8 +41,8 @@ export default function RegionsContainer() {
               </>
             ) : null}
           </button>
-        </li>
+        </MenuItem>
       ))}
-    </ul>
+    </MenuList>
   );
 }
