+++ src/RestaurantsContainer.jsx
@@ -1,3 +1,5 @@
+import React from 'react';
+
 import { useSelector } from 'react-redux';
 
 import { get } from './utils';
